requirements.txt
# === BACKEND ===
fastapi==0.110.1
uvicorn[standard]==0.29.0

# === WORKER ===
celery==5.5.2
redis==5.0.4

# === OCR ENGINE ===
ocrmypdf==14.1.0
reportlab>=3.6.0
pdfminer.six==20221105
Pillow>=9.2.0
img2pdf>=0.4.4
coloredlogs>=15.0.1
python-docx>=0.8.11

# === UTILITAIRES ===
python-multipart==0.0.9
python-dotenv==1.0.1

# === LOGGING ===
loguru>=0.7.0

# === MODELS ===
pydantic==1.10.12
typing-extensions>=4.7.1
